Organized project layout by grouping interfaces w/ their implementations and corresponding testers
