+++ src/Bag/LinkedBag.java
@@ -1,6 +1,4 @@
-package APIs;
-
-import Stack.LinkedStack;
+package Bag;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
